refactor: make ActionForm more responsible for code reusability
