+++ components/produce.tsx
@@ -1,18 +1,9 @@
-import {gql, useMutation} from "@apollo/client";
-import React, {useState} from "react";
-import {useAuth} from "../lib/auth";
+import {gql} from "@apollo/client";
+import React from "react";
 import ActionForm from "./ActionForm";
 
 const Produce = (props:{processId:string}) => {
 
-    const [unitId, setUnitId] = useState('')
-    const [quantity, setQuantity] = useState(0)
-    const [resourceType, setResourceType] = useState('')
-    const [hasPointInTime, setHasPointInTime] = useState('')
-    const [resourceName, setResourceName] = useState('')
-    const [resourceNote, setResourceNote] = useState('')
-
-    const {authId} = useAuth()
     const PRODUCE_MUTATION = gql`
             mutation (
               $outputOf: ID!
@@ -73,37 +64,8 @@ const Produce = (props:{processId:string}) => {
             }
           `
 
-    const [result, { data, loading, error }] = useMutation(PRODUCE_MUTATION)
-
-  function onSubmit(e:any) {
-    result({variables:{provider: authId,
-                              receiver: authId,
-                              outputOf: props.processId,
-                              newInventoriedResource: { name:resourceName, note: resourceNote},
-                              resourceConformsTo: resourceType,
-                              resourceQuantity: {hasNumericalValue:quantity,hasUnit:unitId },
-                              hasPointInTime: new Date(hasPointInTime).toISOString()}})
-    e.preventDefault()
-  }
-   const handleUnit = (e:React.ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault()
-    setUnitId(e.target.value)
-  }
-   const handleResource = (e:React.ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault()
-    setResourceType(e.target.value)
-  }
-
   return (
-      <ActionForm handleUnit={handleUnit}
-                  onSubmit={onSubmit}
-                  setQuantity={(e:number) => setQuantity(e)}
-                  setResourceName={(e:string)=>setResourceName(e)}
-                  handleResource={handleResource}
-                  setResourceNote={(e:string)=>setResourceNote(e)}
-                  setHasPointInTime={(e:string)=>setHasPointInTime(e)}
-                  type="produce"
-      />
+      <ActionForm MUTATION={PRODUCE_MUTATION} processId={props.processId} type="produce"/>
   )};
 
 export default Produce
