+++ components/raise.tsx
@@ -1,19 +1,9 @@
-import {gql, useMutation} from "@apollo/client";
-import React, {useState} from "react";
-import SelectResourceType from "../components/select_resource_type";
-import SelectUnit from "../components/select_unit";
-import {useAuth} from "../lib/auth";
+import {gql} from "@apollo/client";
+import React from "react";
+import ActionForm from "./ActionForm";
 
 const Raise = (props:{processId:string}) => {
 
-    const [unitId, setUnitId] = useState('')
-    const [quantity, setQuantity] = useState(0)
-    const [resourceType, setResourceType] = useState('')
-    const [hasPointInTime, setHasPointInTime] = useState('')
-    const [resourceName, setResourceName] = useState('')
-    const [resourceNote, setResourceNote] = useState('')
-
-    const {authId} = useAuth()
     const RAISE_MUTATION = gql`
             mutation (
               $outputOf: ID!
@@ -74,43 +64,12 @@ const Raise = (props:{processId:string}) => {
             }
           `
 
-    const [result, { data, loading, error }] = useMutation(RAISE_MUTATION)
-
-  function onSubmit(e:any) {
-    result({variables:{provider: authId,
-                              receiver: authId,
-                              outputOf: props.processId,
-                              newInventoriedResource: { name:resourceName, note: resourceNote},
-                              resourceConformsTo: resourceType,
-                              resourceQuantity: {hasNumericalValue:quantity,hasUnit:unitId },
-                              hasPointInTime: new Date(hasPointInTime).toISOString()}})
-    e.preventDefault()
-  }
-   const handleUnit = (e:React.ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault()
-    setUnitId(e.target.value)
-  }
-   const handleResource = (e:React.ChangeEvent<HTMLInputElement>) => {
-    e.preventDefault()
-    setResourceType(e.target.value)
-  }
 
   return (
-      <form onSubmit={onSubmit}>
-          <SelectResourceType handleSelect={handleResource}/>
-          <input type="number"
-                 placeholder="Type here"
-                 className="input input-bordered"
-                 onChange={(e) => setQuantity(parseInt(e.target.value))}
-          />
-          <SelectUnit handleSelect={handleUnit}/>
-          <textarea onChange={(e)=>setResourceName(e.target.value)} className="textarea textarea-bordered w-full" placeholder="Resource Descrption"/>
-          <textarea onChange={(e)=>setResourceNote(e.target.value)} className="textarea textarea-bordered w-full" placeholder="Note"/>
-          <input onChange={(e)=>setHasPointInTime(e.target.value)} type="date" placeholder="Date 1" className="input input-bordered"/>
-          <input type="date" placeholder="Date 2" className="input input-bordered"/>
-          <input type="date" placeholder="Date 3" className="input input-bordered"/>
-          <button type="submit" className="btn btn-primary float-right">Raise</button>
-      </form>
+      <ActionForm MUTATION={RAISE_MUTATION}
+                  processId={props.processId}
+                  type="raise"
+      />
   )};
 
 export default Raise
